Move squid.db.utils to squid.utils
